Move REQUEST-STATUS properties to new type system
