+++ ical/types/__init__.py
@@ -9,7 +9,17 @@ from .cal_address import CalAddress
 from .geo import Geo
 from .period import Period
 from .recur import Recur
+from .request_status import RequestStatus
 from .uri import Uri
 from .utc_offset import UtcOffset
 
-__all__ = ["const", "CalAddress", "Geo", "Period", "Recur", "UtcOffset", "Uri"]
+__all__ = [
+    "const",
+    "CalAddress",
+    "Geo",
+    "Period",
+    "Recur",
+    "RequestStatus",
+    "UtcOffset",
+    "Uri",
+]
